extract logger, redispubsub; export all
