+++ bigbluebutton-html5/server/helpers.js
@@ -28,3 +28,10 @@ export function clearCollections() {
   clearPollCollection();
   clearCursorCollection();
 }
+
+
+export const bind = function (fn, me) { return function () { return fn.apply(me, arguments); }; };
+
+export const indexOf = [].indexOf || function (item) {
+    for (let i = 0, l = this.length; i < l; i++) { if (i in this && this[i] === item) return i; } return -1;
+  };
